alex/react-svg-exports: remove unused bits
