+++ packages/tldraw/src/lib/shapes/shared/getBrowserCanvasMaxSize.tsx
@@ -8,7 +8,7 @@ export type CanvasMaxSize = {
 
 let maxSizePromise: Promise<CanvasMaxSize> | null = null
 
-export function getBrowserCanvasMaxSize() {
+function getBrowserCanvasMaxSize() {
 	if (!maxSizePromise) {
 		maxSizePromise = calculateBrowserCanvasMaxSize()
 	}
@@ -28,8 +28,8 @@ async function calculateBrowserCanvasMaxSize(): Promise<CanvasMaxSize> {
 }
 
 // https://github.com/jhildenbiddle/canvas-size?tab=readme-ov-file#test-results
-export const MAX_SAFE_CANVAS_DIMENSION = 8192
-export const MAX_SAFE_CANVAS_AREA = 4096 * 4096
+const MAX_SAFE_CANVAS_DIMENSION = 8192
+const MAX_SAFE_CANVAS_AREA = 4096 * 4096
 
 export async function clampToBrowserMaxCanvasSize(width: number, height: number) {
 	if (
