--- packages/validate/src/index.d.ts
@@ -1,10 +1,0 @@
-import * as T from './lib/validation'
-export {
-	ArrayOfValidator,
-	DictValidator,
-	ObjectValidator,
-	UnionValidator,
-	Validator,
-} from './lib/validation'
-export { T }
-//# sourceMappingURL=index.d.ts.map
