--- packages/validate/src/test/validation.test.d.ts
@@ -1,2 +1,0 @@
-export {}
-//# sourceMappingURL=validation.test.d.ts.map
